dataset_creation/requirements.txt
openai==1.69.0
python-dotenv
pandas
fastparquet
huggingface_hub
datasets
opencv-python